Expose Duplicates

Sometimes datasets are expected to be tidy but aren't, finding distinct rows of duplicated IDs is easy but finding why they're distinct in many column tables is less straight forward. This functions returns the values that resulted in any duplicated IDs in one of two forms either a named list or a tibble

exposeDupes(x, grouping_var, listout = TRUE)

Arguments

x
Tibble or Dataframe

grouping_var
Column to look for duplicated values

listout
Flag to return either list or tibble

Value

List or Dataframe of results

Details

Named list of two-column tibbles for each value resulting in duplicate IDs

Grouping Variable

Distinct values

Tibble with the following columns

Grouping Variable

n * X.grpNdistinct number of distinct values for duplicated ID

n * X.values values for that duplicated ID

https://dplyr.tidyverse.org/articles/programming.html

Examples

# \donttest{
  df <- data.frame(name = sample(letters, 20, replace = TRUE),
               month = sample(month.name, 20, replace = TRUE),
            letters = sample(LETTERS[1:10], 20, replace = TRUE),
            nums = floor(runif(20, 1, 15)))
  dplyr::count(df, name)

#>    name n
#> 1     b 1
#> 2     d 1
#> 3     e 4
#> 4     f 2
#> 5     i 1
#> 6     k 2
#> 7     l 1
#> 8     m 2
#> 9     o 3
#> 10    w 2
#> 11    x 1

  exposeDupes(df, name)

#> $letters
#> # A tibble: 12 x 2
#>    name  letters
#>    <chr> <chr>  
#>  1 e     F      
#>  2 e     H      
#>  3 e     A      
#>  4 k     J      
#>  5 k     C      
#>  6 m     H      
#>  7 m     C      
#>  8 o     C      
#>  9 o     A      
#> 10 o     J      
#> 11 w     F      
#> 12 w     G      
#> 
#> $month
#> # A tibble: 15 x 2
#>    name  month    
#>    <chr> <chr>    
#>  1 e     June     
#>  2 e     July     
#>  3 e     May      
#>  4 e     April    
#>  5 f     April    
#>  6 f     June     
#>  7 k     February 
#>  8 k     January  
#>  9 m     May      
#> 10 m     June     
#> 11 o     February 
#> 12 o     March    
#> 13 o     September
#> 14 w     December 
#> 15 w     May      
#> 
#> $nums
#> # A tibble: 15 x 2
#>    name  nums 
#>    <chr> <chr>
#>  1 e     7    
#>  2 e     3    
#>  3 e     8    
#>  4 e     12   
#>  5 f     13   
#>  6 f     1    
#>  7 k     1    
#>  8 k     7    
#>  9 m     9    
#> 10 m     4    
#> 11 o     6    
#> 12 o     3    
#> 13 o     11   
#> 14 w     4    
#> 15 w     12   
#> 

  exposeDupes(df, name, listout = FALSE)

#> # A tibble: 15 x 7
#>    name  letters.grpNdistinct letters.values month.grpNdistinct month.values
#>    <chr>                <int> <chr>                       <int> <chr>       
#>  1 e                        3 F                               4 June        
#>  2 e                        3 H                               4 July        
#>  3 e                        3 A                               4 May         
#>  4 e                        3 F                               4 April       
#>  5 f                        1 B                               2 April       
#>  6 f                        1 B                               2 June        
#>  7 k                        2 J                               2 February    
#>  8 k                        2 C                               2 January     
#>  9 m                        2 H                               2 May         
#> 10 m                        2 C                               2 June        
#> 11 o                        3 C                               3 February    
#> 12 o                        3 A                               3 March       
#> 13 o                        3 J                               3 September   
#> 14 w                        2 F                               2 December    
#> 15 w                        2 G                               2 May         
#> # ... with 2 more variables: nums.grpNdistinct <int>, nums.values <dbl>

  # }